requirements.txt
fastapi
uvicorn[standard]
anthropic
pyyaml
python-dotenv
httpx
